+++ demoproj/urls.py
@@ -15,8 +15,9 @@ Including another URLconf
 """
 from django.urls import path
 
-from demoproj.views import index_view
+from demoproj.views import index_view, rest_view
 
 urlpatterns = [
     path('', index_view, name='index'),
+    path('api', rest_view, name='drf'),
 ]