+++ django_guid/config.py
@@ -13,6 +13,7 @@ class Settings(object):
         self.GUID_HEADER_NAME = 'Correlation-ID'
         self.VALIDATE_GUID = True
         self.SKIP_CLEANUP = False
+        self.RETURN_HEADER = True
 
         if hasattr(django_settings, 'DJANGO_GUID'):
             _settings = django_settings.DJANGO_GUID
@@ -30,6 +31,8 @@ class Settings(object):
                 raise ImproperlyConfigured('VALIDATE_GUID must be a boolean')
             if not isinstance(self.GUID_HEADER_NAME, str):
                 raise ImproperlyConfigured('GUID_HEADER_NAME must be a string')  # Note: Case insensitive
+            if not isinstance(self.RETURN_HEADER, bool):
+                raise ImproperlyConfigured('RETURN_HEADER must be a boolean')
         else:
             pass  # Do nothing if DJANGO_GUID not found in settings