+++ tests/unit/test_config.py
@@ -29,12 +29,24 @@ def test_invalid_skip_guid_setting(monkeypatch):
         Settings()
 
 
+def test_invalid_return_header_setting(monkeypatch):
+    monkeypatch.setattr(django_settings, 'DJANGO_GUID', {'RETURN_HEADER': 'string'})
+    with pytest.raises(ImproperlyConfigured, match='RETURN_HEADER must be a boolean'):
+        Settings()
+
+
 def test_valid_settings(monkeypatch):
     monkeypatch.setattr(
         django_settings,
         'DJANGO_GUID',
-        {'SKIP_CLEANUP': True, 'VALIDATE_GUID': False, 'GUID_HEADER_NAME': 'Correlation-ID-TEST'},
+        {
+            'SKIP_CLEANUP': True,
+            'VALIDATE_GUID': False,
+            'GUID_HEADER_NAME': 'Correlation-ID-TEST',
+            'RETURN_HEADER': False,
+        },
     )
     assert not Settings().VALIDATE_GUID
     assert Settings().SKIP_CLEANUP
     assert Settings().GUID_HEADER_NAME == 'Correlation-ID-TEST'
+    assert not Settings().RETURN_HEADER